# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: CasaRoca System AI · API
  version: "1.0.0"
  description: |
    API del sistema eclesial de Casa Sobre la Roca: una central y 36 sedes.

    ESTE ARCHIVO SE GENERA. No lo edite a mano: corra `scripts/generar-openapi.js`.
    La verificacion de entrega compara lo generado con lo versionado.

    Autenticacion: token de portador (Bearer). El token dice que sesion DICE ser;
    quien decide es la base, que en cada peticion responde si esa sesion sigue viva.
    Por eso cerrar una sesion surte efecto en el instante.
servers:
  - url: https://api.casaroca.org
    description: produccion
  - url: https://api.staging.casaroca.org
    description: pruebas
  - url: http://127.0.0.1:3000
    description: desarrollo
components:
  securitySchemes:
    portador:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error: { type: boolean }
        mensaje: { type: string, description: "En castellano y con la accion siguiente. Nunca filtra la estructura interna." }
        peticionId: { type: string, description: "Identificador de traza. El usuario lo lee por telefono al soporte." }
security:
  - portador: []
paths:
...
  /api/v1/grupos/:id/miembros/:personaId/salir:
    post:
      summary: "Sacar a alguien"
      description: "EXIGE motivo: «se fue» no le dice nada al que viene detras."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/v1/grupos/:id/miembros:
    post:
      summary: "Agregar a alguien"
      description: "Idempotente: si ya esta activo, lo dice y no duplica."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/v1/grupos/:id/reuniones:
    post:
      summary: "Reportar una reunion"
      description: "Es lo que convierte una lista en un grupo vivo. Una por grupo y dia."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/v1/grupos/:id:
    get:
      summary: "La ficha del grupo"
      description: "Quien esta, quien se fue y con que motivo, y las ultimas veinticuatro reuniones."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/v1/grupos:
    get:
      summary: "Los grupos"
      description: "Ordenados por el que lleva mas tiempo sin reportar reunion: lo que hay que mirar primero no es el grupo grande, es el callado."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    post:
      summary: "Crear un grupo"
      description: "Tipo, dia, hora y cupo. El ministerio y el segmento son opcionales."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...
  /api/v1/rocakids/checkin:
    post:
      summary: "Registrar la entrada de un menor"
      description: "Devuelve el codigo UNA vez; despues solo existe cifrado. IDEMPOTENTE por menor, sala y dia: la cola sin conexion puede reintentar sin duplicar el ingreso."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/v1/rocakids/entregar:
    post:
      summary: "Entregar al menor"
      description: "La base verifica acudiente autorizado Y codigo. Un intento fallido queda registrado con hora y nombre."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/v1/rocakids/menores/:id/acudientes:
    get:
      summary: "Quien puede entregar y retirar a este menor"
      description: "La pantalla ofrece esta lista, no un campo de texto libre: un texto libre convierte la salvaguarda en una formalidad."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/v1/rocakids/salas/:id/entrar-a-servir:
    post:
      summary: "Entrar a servir en una sala"
      description: "La base rechaza a quien no tenga antecedentes de salvaguarda vigentes."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/v1/rocakids/salas/:id/roster:
    get:
      summary: "Censo de la sala"
      description: "Quien puede entrar hoy: nombre, edad y si ya esta dentro. NO devuelve condiciones medicas ni codigos: esto se guarda en el equipo para trabajar sin conexion, y lo que se guarda en un equipo se pierde con el equipo."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/v1/rocakids/salas/:id/servidores:
    get:
      summary: "Quien sirve hoy en la sala"
      description: "Es la respuesta a «quien estaba con los ninos ese dia», la primera pregunta de cualquier incidente."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/v1/rocakids/salas:
    get:
      summary: "Salas de ninos y su estado de ahora"
      description: "Con cuantos ninos hay dentro y si se cumple la regla de dos adultos. Exige N4."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...